+++ packages/preview-middleware-client/test/unit/cpe/changes/service.test.ts
@@ -1792,7 +1792,7 @@ describe('ChangeService', () => {
             value: 'abc'
         });
 
-        expect(sendActionMock).toBeCalledWith(
+        expect(sendActionMock).toHaveBeenCalledWith(
             propertyChangeFailed({
                 controlId: 'control1',
                 propertyName: 'text',