+++ packages/flp-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sap-ux/flp-config-sub-generator
 
+## 0.3.1
+
+### Patch Changes
+
+-   @sap-ux/flp-config-inquirer@0.4.1
+-   @sap-ux/app-config-writer@0.6.23
+
+## 0.3.0
+
+### Minor Changes
+
+-   089b56f: updates extension types and extends prompt options in sub gens
+
+### Patch Changes
+
+-   Updated dependencies [089b56f]
+    -   @sap-ux/flp-config-inquirer@0.4.0
+
 ## 0.2.55
 
 ### Patch Changes