+++ packages/ui-service-inquirer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/ui-service-inquirer
 
+## 0.1.50
+
+### Patch Changes
+
+-   Updated dependencies [089b56f]
+    -   @sap-ux/abap-deploy-config-inquirer@1.5.0
+
 ## 0.1.49
 
 ### Patch Changes